Implement for MatRef
Also uses the trick from sarah to work for any type Rows/Cols
