+++ faer/src/mat/matown.rs
@@ -1439,91 +1439,6 @@ impl<T, Rows: Shape> Mat<T, Rows, usize> {
 	}
 }
 
-impl<T> Mat<T, usize, usize>
-where
-	T: PartialOrd + Clone,
-{
-	/// Returns the maximum element in the matrix
-	///
-	/// # Returns
-	///
-	/// * `Option<T>` - The maximum element in the matrix, or `None` if the matrix is empty
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use faer::{mat, Mat};
-	///
-	/// let m = mat![
-	///     [1.0, 5.0, 3.0],
-	///     [4.0, 2.0, 9.0],
-	///     [7.0, 8.0, 6.0],
-	/// ];
-	///
-	/// assert_eq!(m.max(), Some(9.0));
-	///
-	/// let empty: Mat<f64> = Mat::new();
-	/// assert_eq!(empty.max(), None);
-	/// ```
-	pub fn max(&self) -> Option<T> {
-		if self.nrows().unbound() == 0 || self.ncols().unbound() == 0 {
-			return None;
-		}
-
-		let mut max_val = self.get(0, 0);
-
-		self.row_iter().for_each(|row| {
-			row.iter().for_each(|val| {
-				if val > max_val {
-					max_val = &val;
-				}
-			});
-		});
-
-		Some((*max_val).clone())
-	}
-
-	/// Returns the minimum element in the matrix
-	///
-	/// # Returns
-	///
-	/// * `Option<T>` - The minimum element in the matrix, or `None` if the matrix is empty
-	///
-	/// # Examples
-	///
-	/// ```
-	/// use faer::{mat, Mat};
-	///
-	/// let m = mat![
-	///     [1.0, 5.0, 3.0],
-	///     [4.0, 2.0, 9.0],
-	///     [7.0, 8.0, 6.0],
-	/// ];
-	///
-	/// assert_eq!(m.min(), Some(1.0));
-	///
-	/// let empty: Mat<f64> = Mat::new();
-	/// assert_eq!(empty.min(), None);
-	/// ```
-	pub fn min(&self) -> Option<T> {
-		if self.nrows().unbound() == 0 || self.ncols().unbound() == 0 {
-			return None;
-		}
-
-		let mut min_val = self.get(0, 0);
-
-		self.row_iter().for_each(|row| {
-			row.iter().for_each(|val| {
-				if val < min_val {
-					min_val = &val;
-				}
-			});
-		});
-
-		Some((*min_val).clone())
-	}
-}
-
 #[cfg(test)]
 mod tests {
 	use crate::{assert, mat};
